feat: make Aliucord installation update button work
Launch a patch opts screen with prefilled options based on the previous install.

+++ app/src/main/kotlin/com/aliucord/manager/util/ColorSerializer.kt
@@ -23,6 +23,6 @@ object ColorSerializer : KSerializer<Color> {
 
     override fun deserialize(decoder: Decoder): Color {
         val string = decoder.decodeString()
-        return Color(string.toInt(16))
+        return Color(string.toLong(16))
     }
 }
